requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
drf-spectacular==0.27.2
django-python3-ldap==0.15.7
# python-csv==0.0.13